Back/requirements.txt
openai<1.0.0
langchain
langchain-experimental
tabulate
qvd
flask
python-dotenv
pandas
datetime